+++ dflib/src/test/java/com/nhl/dflib/DataFrame_FilterTest.java
@@ -26,4 +26,16 @@ public class DataFrame_FilterTest {
                 .expectHeight(1)
                 .expectRow(0, 20);
     }
+
+    @Test
+    public void testFilter_WithBooleanSeries() {
+
+        DataFrame df = DataFrame.newFrame("a").foldByRow(10, 20, 30)
+                .filterRows(BooleanSeries.forBooleans(true, false, true));
+
+        new DataFrameAsserts(df, "a")
+                .expectHeight(2)
+                .expectRow(0, 10)
+                .expectRow(1, 30);
+    }
 }