+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #119 Primitive (custom) accumulators for DataFrameBuilder.byRow(..)
 * #120 DataFrame.pivot(..)
 * #121 BigDecimal aggregators with controlled scale
+* #123 Improve performance of "DF.filterRows(BooleanSeries)" 
 
 ## Release 0.10